finished desktop and mobile layout
