+++ frontend/src/components/ui/NavBar/NavBar.css
@@ -1,0 +1,69 @@
+* {
+  color: #261c1c;
+}
+
+.nav-bar {
+  box-sizing: border-box;
+  margin: 0;
+  width: 100dvw;
+  padding: 14px 10%;
+  display: flex;
+  flex-wrap: wrap;
+  justify-content: space-between;
+  align-items: center;
+  overflow-x: hidden;
+  background-color: #add9d9;
+}
+
+.name {
+  padding: 0;
+  /* float: left; */
+  margin-right: 6px;
+}
+
+ul {
+  /* float: right; */
+  list-style-type: none;
+  display: flex;
+  justify-content: center;
+  max-width: 80dvw;
+  gap: 24px;
+  margin: 0;
+  padding: 0;
+}
+
+li {
+  font-size: 16px;
+  text-align: center;
+  text-wrap: nowrap;
+  margin: 0;
+  padding: 0;
+}
+
+.link {
+  text-decoration: none;
+}
+
+.btn-menu {
+  display: none;
+}
+
+@media screen and (max-width: 600px) {
+  ul {
+    display: none;
+  }
+  li {
+    display: none;
+  }
+  .btn-menu {
+    display: contents;
+    margin: 0;
+    padding: 0;
+  }
+  .active-item {
+    display: block;
+    box-sizing: border-box;
+    width: 100dvw;
+    text-align: left;
+  }
+}
